+++ tests/Jroc.Tests/Import/JavaScript/Import_ReexportAsDefault_Reexport.mjs
@@ -1,0 +1,3 @@
+"use strict";
+
+export { payload as default } from "./Import_ReexportAsDefault_Lib.mjs";